refactor(types): 将 MCP 相关类型定义提取到独立文件
将 MCP 服务器配置相关的 Zod schema 和类型定义从 index.ts 移动到新的 mcp.ts 文件
保持原有功能不变，提高代码组织性和可维护性
